LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/forward-already-detached-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back


Harness Error (TIMEOUT), message = null

TIMEOUT forward() in a detached window Test timed out
